Made bots use aoe even if they have a CC mark set
Since all companions come with moon as standart cc mark I had to remove this check. But lets be honest, it is not only the worst check for breakable ccs, it also made them not use aoe, EVEN TAUNTS, at all after setting up a CC mark. I created a new check for this and added it to all damage ae spells. They still break CCs sometimes, but now it is so much better than before.